requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
docxtpl==0.18.0
python-docx==1.1.2
pydantic==2.9.2
pypinyin==0.53.0
mcp==1.2.0
